requirements.txt
streamlit
pycryptodome
